+++ server/src/main/kotlin/org/kryptonmc/krypton/entity/metadata/MetadataKey.kt
@@ -23,7 +23,7 @@ import org.kryptonmc.internal.annotations.CataloguedBy
  * A key that can be used to retrieve entity metadata from a [MetadataHolder].
  */
 @CataloguedBy(MetadataKeys::class)
-class MetadataKey<T>(val id: Int, val serializer: MetadataSerializer<T>) {
+class MetadataKey<T>(val id: Byte, val serializer: MetadataSerializer<T>) {
 
     override fun toString(): String = "MetadataKey(id=$id, serializer=$serializer)"
 }
